+++ packages/firebase_database/ios/Classes/FirebaseDatabasePlugin.m
@@ -1,3 +1,7 @@
+// Copyright 2017 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 #import "FirebaseDatabasePlugin.h"
 
 #import <Firebase/Firebase.h>
@@ -19,6 +23,42 @@ FIRDatabaseReference *getReference(NSDictionary *arguments) {
   FIRDatabaseReference *ref = [FIRDatabase database].reference;
   if ([path length] > 0) ref = [ref child:path];
   return ref;
+}
+
+FIRDatabaseQuery *getQuery(NSDictionary *arguments) {
+  FIRDatabaseQuery *query = getReference(arguments);
+  NSDictionary *parameters = arguments[@"parameters"];
+  NSString *orderBy = parameters[@"orderBy"];
+  if ([orderBy isEqualToString:@"child"]) {
+    query = [query queryOrderedByChild:parameters[@"orderByChildKey"]];
+  } else if ([orderBy isEqualToString:@"key"]) {
+    query = [query queryOrderedByKey];
+  } else if ([orderBy isEqualToString:@"value"]) {
+    query = [query queryOrderedByValue];
+  } else if ([orderBy isEqualToString:@"priority"]) {
+    query = [query queryOrderedByPriority];
+  }
+  id startAt = parameters[@"startAt"];
+  if (startAt) {
+    query = [query queryStartingAtValue:startAt childKey:parameters[@"endAtKey"]];
+  }
+  id endAt = parameters[@"endAt"];
+  if (endAt) {
+    query = [query queryEndingAtValue:endAt childKey:parameters[@"endAtKey"]];
+  }
+  id equalTo = parameters[@"equalTo"];
+  if (equalTo) {
+    query = [query queryEqualToValue:equalTo];
+  }
+  NSNumber *limitToFirst = parameters[@"limitToFirst"];
+  if (limitToFirst) {
+    query = [query queryLimitedToFirst:limitToFirst.intValue];
+  }
+  NSNumber *limitToLast = parameters[@"limitToLast"];
+  if (limitToLast) {
+    query = [query queryLimitedToLast:limitToLast.intValue];
+  }
+  return query;
 }
 
 FIRDataEventType parseEventType(NSString *eventTypeString) {
@@ -67,12 +107,28 @@ FIRDataEventType parseEventType(NSString *eventTypeString) {
       ^(NSError *error, FIRDatabaseReference *ref) {
         result(error.flutterError);
       };
-  if ([@"DatabaseReference#set" isEqualToString:call.method]) {
+  if ([@"FirebaseDatabase#goOnline" isEqualToString:call.method]) {
+    [[FIRDatabase database] goOnline];
+  } else if ([@"FirebaseDatabase#goOffline" isEqualToString:call.method]) {
+    [[FIRDatabase database] goOffline];
+  } else if ([@"FirebaseDatabase#purgeOutstandingWrites" isEqualToString:call.method]) {
+    [[FIRDatabase database] purgeOutstandingWrites];
+  } else if ([@"FirebaseDatabase#setPersistenceEnabled" isEqualToString:call.method]) {
+    NSNumber *value = call.arguments[@"value"];
+    [FIRDatabase database].persistenceEnabled = value.boolValue;
+  } else if ([@"FirebaseDatabase#setPersistenceCacheSizeBytes" isEqualToString:call.method]) {
+    NSNumber *value = call.arguments[@"value"];
+    [FIRDatabase database].persistenceCacheSizeBytes = value.unsignedIntegerValue;
+  } else if ([@"DatabaseReference#set" isEqualToString:call.method]) {
     [getReference(call.arguments) setValue:call.arguments[@"value"]
+                               andPriority:call.arguments[@"priority"]
                        withCompletionBlock:defaultCompletionBlock];
+  } else if ([@"DatabaseReference#setPriority" isEqualToString:call.method]) {
+    [getReference(call.arguments) setPriority:call.arguments[@"priority"]
+                          withCompletionBlock:defaultCompletionBlock];
   } else if ([@"Query#observe" isEqualToString:call.method]) {
     FIRDataEventType eventType = parseEventType(call.arguments[@"eventType"]);
-    __block FIRDatabaseHandle handle = [getReference(call.arguments)
+    __block FIRDatabaseHandle handle = [getQuery(call.arguments)
                       observeEventType:eventType
         andPreviousSiblingKeyWithBlock:^(FIRDataSnapshot *snapshot, NSString *previousSiblingKey) {
           [self.channel invokeMethod:@"Event"
@@ -88,8 +144,11 @@ FIRDataEventType parseEventType(NSString *eventTypeString) {
     result([NSNumber numberWithUnsignedInteger:handle]);
   } else if ([@"Query#removeObserver" isEqualToString:call.method]) {
     FIRDatabaseHandle handle = [call.arguments[@"handle"] unsignedIntegerValue];
-    [getReference(call.arguments) removeObserverWithHandle:handle];
+    [getQuery(call.arguments) removeObserverWithHandle:handle];
     result(nil);
+  } else if ([@"Query#keepSynced" isEqualToString:call.method]) {
+    NSNumber *value = call.arguments[@"value"];
+    [getQuery(call.arguments) keepSynced:value];
   } else {
     result(FlutterMethodNotImplemented);
   }