+++ packages/google_sign_in/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* Fix signOut and disconnect (they were silently ignored)
+* Fix test (#10050)
+
 ## 0.0.2
 
 * Don't try to sign in again if user is already signed in